fix: resolve all 7 mypy type annotation errors (zero runtime changes)

+++ phronesisml/ml/preflight/memory.py
@@ -110,7 +110,7 @@ class MemorySafety:
             import psutil
 
             mem = psutil.virtual_memory()
-            return mem.total / (1024**3)
+            return float(mem.total / (1024**3))
         except Exception:
             return 8.0
